+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 
 ### Added
+- Added debug log emits for the ``DataFrameETL`` transformer (#24, #27)
 - Emit a warning if the user attempts to expand a column with
   too many categories (#25, #26)